feat: Make `delete` command a separate package named `remove`
