+++ cmd/remove/command.go
@@ -1,15 +1,13 @@
-package cmd
+package remove
 
 import (
-	"yasm/scripts"
-
 	"github.com/urfave/cli/v2"
 )
 
 // deleteCommand returns a new CLI command for deleting scripts.
 // This command allows the user to delete an existing script by name,
 // or open a menu to select and delete a script if no name is provided.
-func deleteCommand() *cli.Command {
+func Command() *cli.Command {
 	return &cli.Command{
 		Name:  "delete",
 		Usage: "Delete an existing script",
@@ -19,7 +17,7 @@ func deleteCommand() *cli.Command {
 
 		Action: func(c *cli.Context) error {
 			scriptName := c.Args().First()
-			return scripts.DeleteScript(scriptName)
+			return removeScript(scriptName)
 		},
 	}
 }
